Reorganized dba source
Rearrange the dba source to make the structure of multiple different
databases (one for resources, one for users) separate.
