+++ dba/fieldmap/doc.go
@@ -1,0 +1,3 @@
+// Package map contains all the necessary code to set up Bleve document
+// mappings.
+package fieldmap
